milvus/requirements.txt
pymilvus==2.6.10
python-dotenv>=1.0.1
pyyaml>=6.0.2
langchain-ollama==1.0.1
ollama==0.6.1